hold the shader bytecode and the D3D out parameters in DWORD
DWORD is unsigned long on Windows and uint32_t everywhere else, which is why
these compiled for twenty years and stop now. Two different faults share the
cause.

The out parameters are only a type mismatch: CheckDeviceMultiSampleType and
ValidateDevice want DWORD*, and were handed the address of an unsigned long.
Five locals in Gfx.cpp and GfxRender.cpp.

The shader bytecode is a layout fault, and a quieter one. GfxShaders.cpp holds
128 arrays of D3D9 shader tokens, declared unsigned long. A token is 32 bits.
Where unsigned long is 64 bits each token occupies eight bytes, so the array is
twice the length it should be and every token after the first sits at the wrong
offset: not a stream D3D9 could reject cleanly, a stream that is not one. The
declarations in GfxShadersDescr.h and the two pointers in GfxRender.cpp that
walk them move with it.

Windows is unaffected in both cases, DWORD being what these already were there.

+++ Versions/Temporary/Engine/Sources/3Dmotor/Gfx.cpp
@@ -58,7 +58,7 @@ static int nDeviceCreationID = 1;
 static SSystemInfo systemInfo;
 static bool bIsDebugRuntime;
 static bool bCanDoFastScreenshot = false;
-static unsigned long nFSAA, nMaxFSAA;
+static DWORD nFSAA, nMaxFSAA;
 static int nWinXPos = 0;
 static int nAdapterToUse = D3DADAPTER_DEFAULT;
 
@@ -363,7 +363,7 @@ static void FillFSAA( D3DPRESENT_PARAMETERS *pRes )
 {
 	if ( nFSAA > 0 && nMaxFSAA > 0 )
 	{
-		unsigned long nFront, nDepth;
+		DWORD nFront, nDepth;
 		if ( 
 			SUCCEEDED( pD3D->CheckDeviceMultiSampleType( GetAdapterToUse(), DEVICE_TYPE,
 				pRes->BackBufferFormat, pRes->Windowed, D3DMULTISAMPLE_NONMASKABLE, &nFront ) ) &&
@@ -629,7 +629,7 @@ static bool CheckDeviceCaps()
 	b8888Supported = 
 		SUCCEEDED( pD3D->CheckDeviceFormat( GetAdapterToUse(), DEVICE_TYPE, desktop.Format, 0, D3DRTYPE_TEXTURE, D3DFMT_A8R8G8B8 ) );
 
-	unsigned long nFrontFSAA, nDepthFSAA;
+	DWORD nFrontFSAA, nDepthFSAA;
 	if ( 
 		SUCCEEDED( pD3D->CheckDeviceMultiSampleType( GetAdapterToUse(), DEVICE_TYPE, D3DFMT_X8R8G8B8,
 			FALSE, D3DMULTISAMPLE_NONMASKABLE, &nFrontFSAA ) ) &&
